Add secret "null" cipher for unit tests
Having the "null" cipher makes it easy to test the encryption modes even
when LTC was compiled without support for any cipher (for whatever reason).

The downside is that no online service seems to propose this cipher to
encrypt data (no wonder...), therefore it's hard to come up with valid
test vectors. Fortunately, the encryption modes are well documented,
making it easy to compute test vectors by hand.