Improve null support for the tests
Some of the modes require specific block sizes (word-sized or 128-bit blocks).
Sometimes, it also depends whether LTC was compiled with LTC_FAST set or not.

To workaround this requirements, this commit adds two new null ciphers:
* TOMCRYPT_CIPHER_NULL_FAST (word-aligned block cipher)
* TOMCRYPT_CIPHER_NULL_128  (128-bit block cipher)

This comes in addition to the existing TOMCRYPT_CIPHER_NULL 8-bit block cipher.